misc/HtmlTestSuite/results/test-1022.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <meta charset="utf-8"/>
    <title>HTML Test: The style information must be applied to the environment specified by the media attribute</title>
    <link href="http://www.intel.com/" rel="author" title="Intel"/>
    <link href="https://html.spec.whatwg.org/multipage/#attr-style-media" rel="help"/>
    <link href="https://html.spec.whatwg.org/multipage/#the-style-element" rel="help"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
    <style>
      #test {
        width: 100px;
      }
    </style>
    <style id="style">
      #test {
        width: 50px;
      }
    </style>
  </head>
  <body>
    <div id="log"/>
    <div id="test"/>
    <script>
      test(function() {
        var testElement = document.getElementById("test");
        var style = document.getElementById("style");
        var width1, width2;

        width1 = window.getComputedStyle(testElement)["width"];
        assert_equals(width1, "50px", "The style should be applied.");

        style.media = "print";
        width2 = window.getComputedStyle(testElement)["width"];
        assert_equals(width2, "100px", "The style should not be applied.");
      }, "The style information must be applied to the environment specified by the media attribute");
    </script>
  </body>
</html>
